feat!: reorganized CLI commands All commands under set, init, get, and remove have been moved under their noun replacements (i.e. init config -> config init).
This is a first step to reserving verbs for only exec command aliases.
