+++ tools/docsgen/markdown.go
@@ -23,7 +23,9 @@ var (
 		schema.FlowfileDefinitionTitle,
 		schema.ConfigDefinitionTitle,
 		schema.WorkspaceDefinitionTitle,
-		schema.TemplateDefinitionTitle,
+		// TODO: fix schema gem bug where the common executable definitions used by the flowfile
+		// and template schemas are non-deterministically generated
+		// schema.TemplateDefinitionTitle,
 	}
 )
